requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.5.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-braces==1.13.0
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-filter==2.0.0
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-simple-history==2.5.1
django-taggit==0.23.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.0
docopt==0.6.2
drf-yasg==1.11.1
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==2.1.0
openapi-codec==1.3.2
pep8==1.7.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.2.0
pylint-django==2.0.4
pylint-plugin-utils==0.4
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
ruamel.yaml==0.15.80
sendgrid==5.6.0
Serializer==0.2.1
simplejson==3.16.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
yapf==0.25.0
